requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
python-jose[cryptography]>=3.3.0
passlib>=1.7.4
bcrypt==3.2.2
pydantic[email-validator]>=2.5.0
python-dotenv>=1.0.0
boto3>=1.34.0
email-validator
python-multipart
slowapi>=0.1.9
apscheduler>=3.10.0

